Extract shuttle-sync-impl crate from shuttle
Move std::sync primitive implementations (mutex, rwlock, condvar,
barrier, once, mpsc, atomic) into shuttle-sync-impl crate. The shuttle
crate becomes a thin facade that re-exports all public types from the
sub-crate via `pub use shuttle_sync_impl::sync::*`.
