+++ shuttle-sync-impl/src/sync/atomic/bool.rs
