+++ shuttle-sync-impl/src/sync/mutex.rs
@@ -1,9 +1,9 @@
-use crate::current;
-use crate::future::batch_semaphore::{BatchSemaphore, Fairness};
-use crate::runtime::task::TaskId;
-use crate::runtime::thread;
 use crate::sync::{LockResult, PoisonError, TryLockError, TryLockResult};
 use crate::sync::{ResourceSignature, ResourceType};
+use shuttle_core::current;
+use shuttle_core::future::batch_semaphore::{BatchSemaphore, Fairness};
+use shuttle_core::runtime::task::TaskId;
+use shuttle_core::runtime::thread;
 use std::cell::RefCell;
 use std::fmt::{Debug, Display};
 use std::ops::{Deref, DerefMut};
@@ -232,14 +232,14 @@ impl<T: Display + ?Sized> Display for MutexGuard<'_, T> {
     }
 }
 
-impl<T> crate::annotations::WithName for &Mutex<T> {
+impl<T> shuttle_core::annotations::WithName for &Mutex<T> {
     fn with_name_and_kind(self, name: Option<&str>, kind: Option<&str>) -> Self {
         (&self.semaphore).with_name_and_kind(name, kind.or(Some("shuttle::sync::Mutex")));
         self
     }
 }
 
-impl<T> crate::annotations::WithName for Mutex<T> {
+impl<T> shuttle_core::annotations::WithName for Mutex<T> {
     fn with_name_and_kind(self, name: Option<&str>, kind: Option<&str>) -> Self {
         (&self).with_name_and_kind(name, kind);
         self
@@ -252,7 +252,7 @@ mod tests {
 
     #[test]
     fn unique_resource_signature_mutex() {
-        crate::check_random(
+        shuttle_schedulers::check_random(
             || {
                 let mutex1 = Mutex::new(0);
                 let mutex2 = Mutex::new(0);
