+++ shuttle/Cargo.toml
@@ -12,7 +12,7 @@ readme = "../README.md"
 [dependencies]
 shuttle-core = { path = "../shuttle-core" }
 shuttle-schedulers = { path = "../shuttle-schedulers" }
-assoc = "0.1.3"
+shuttle-sync-impl = { path = "../shuttle-sync-impl" }
 bitvec = "1.0.1"
 cfg-if = "1.0"
 hex = "0.4.2"
@@ -21,7 +21,6 @@ rand_core = "0.6.4"
 rand = "0.8.6"
 rand_pcg = "0.3.1"
 scoped-tls = "1.0.0"
-smallvec = { version = "1.11.2", features = ["const_new"] }
 tracing = { version = "0.1.36", default-features = false, features = ["std"] }
 corosensei = "0.3.1"
