+++ shuttle/src/sync/mod.rs
@@ -1,30 +1,3 @@
 //! Shuttle's implementation of [`std::sync`].
 
-pub mod atomic;
-mod barrier;
-mod condvar;
-pub mod mpsc;
-mod mutex;
-mod once;
-mod rwlock;
-
-pub use barrier::{Barrier, BarrierWaitResult};
-pub use condvar::{Condvar, WaitTimeoutResult};
-
-pub use mutex::Mutex;
-pub use mutex::MutexGuard;
-
-pub use once::Once;
-pub use once::OnceState;
-
-pub use rwlock::RwLock;
-pub use rwlock::RwLockReadGuard;
-pub use rwlock::RwLockWriteGuard;
-
-pub use std::sync::{LockResult, PoisonError, TryLockError, TryLockResult};
-
-// TODO implement true support for `Arc`
-pub use std::sync::{Arc, Weak};
-
-// Re-export from shuttle-core
-pub(crate) use shuttle_core::sync_types::{ResourceSignature, ResourceType};
+pub use shuttle_sync_impl::sync::*;
